refactor: Moved some of the logic out of grid.ts
Abstracted some token logic from grid.ts into its own token.ts file.

FSA22V2-252
